feat: add security

+++ src/app/api/admin/clinic/[id]/route.js
@@ -1,11 +1,25 @@
 import { NextResponse } from "next/server";
 import { clinicRepo } from "@/app/lib/db/clinicRepo";
 import { userRepo } from "@/app/lib/db/userRepo";
+import authOptions from "@/app/lib/authoption";
+import { getServerSession } from "next-auth";
 
 export async function PUT(request, { params }) {
     const { id } = params;
     const clinic = await request.json();
-            console.log(id,clinic)
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+    }
+    const sessionEmail = session.user.email;
+    const sessionUser = await userRepo.getUserByEmail(sessionEmail);
+    if (!sessionUser) {
+        return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+    if (sessionUser.role !== "admin") {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+    }
+
     try {
         const updatedClinic = await clinicRepo.updateClinic(id, clinic);
 
@@ -27,12 +41,25 @@ export async function PUT(request, { params }) {
 
 export async function DELETE(request, { params }) {
     const { id } = params;
-    const clinic = await request.json();
-    try {        
+    // const clinic = await request.json();
+    const session = await getServerSession(authOptions);
+    if (!session) {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+    }
+    const sessionEmail = session.user.email;
+    const sessionUser = await userRepo.getUserByEmail(sessionEmail);
+    if (!sessionUser) {
+        return NextResponse.json({ message: "User not found" }, { status: 404 });
+    }
+    if (sessionUser.role !== "admin") {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+    }
+
+    try {
         await userRepo.deleteClinicMembers(id);
         await clinicRepo.deleteClinic(id);
 
-        return NextResponse.json({ success: true, message: "Clinic deleted successfully"});
+        return NextResponse.json({ success: true, message: "Clinic deleted successfully" });
     } catch (error) {
         console.error(error);
         return NextResponse.json({ success: false, message: "Clinic delete failed", error: error.message });
